pyproject.toml
[project]
name = "tinytrash"
version = "0.1.0"
description = "TinyTrash - Waste classification with computer vision"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
